feat: make tap reactions universal across chat view and photo mode
The reaction channel moves from the photomode store to the vrm store, the
scene raycast now fires in the main view too (overlay and XR excluded; both
own their pointer handling), and the applier no longer requires photo mode.
Reactions stay relationship-staged everywhere; the restore-after-update
impulse is safe alongside live idle and talking animations.
